Implementando o banco de dados, e trazendo os primeiros dados dinâmicamente para a view

+++ database/migrations/2023_05_18_012734_create_post_table.php
@@ -15,9 +15,9 @@ return new class extends Migration
             $table->id();
             $table->string('title');
             $table->text('content');
-            $table->integer('category_id');
+            $table->unsignedBigInteger('category_id');
             $table->foreign('category_id')->references('id')->on('categories');
-            $table->integer('author_id');
+            $table->unsignedBigInteger('author_id');
             $table->foreign('author_id')->references('id')->on('authors');
             $table->timestamp('published_at');
             $table->timestamps();
