+++ database/migrations/2023_05_18_014828_create_posts_tags_table.php
@@ -13,8 +13,8 @@ return new class extends Migration
     {
         Schema::create('posts_tags', function (Blueprint $table) {
             $table->id();
-            $table->integer('post_id');
-            $table->integer('tag_id');
+            $table->unsignedBigInteger('post_id');
+            $table->unsignedBigInteger('tag_id');
             $table->foreign('post_id')->references('id')->on('posts')->onDelete('cascade');
             $table->foreign('tag_id')->references('id')->on('tags')->onDelete('cascade');
         });
